refactor(extensions)!: drop Extension.systems in favour of install()
`ApplicationSystemBinding.create(app) => System | undefined` was exactly
`install: app => { app.systems.add(system) }`, down to the reverse-order
destruction `SystemRegistry` performs either way. Since `install` landed, both
hooks run at the same moment, with the same argument, one immediately after the
other — two ways to say one thing, and the guide had to explain which to reach
for.

`install` is the more capable of the two: the same closure that registers a
system can undo it through the returned `ExtensionDisposer`, which a binding
never could. A binding returning `undefined` to opt out becomes an ordinary
`if` inside the installer.

The type had no users outside `src/` and its own tests — no package, example,
guide or template ever declared one. The tests that covered per-application
instances, registration on `app.systems`, reverse-order teardown and a system
whose `destroy()` throws move over to `install` and join the lifetime suite;
`materializeApplicationSystems` and its unit test go with the binding.

BREAKING CHANGE: `Extension.systems` and the `ApplicationSystemBinding` type
are removed. Register app-level systems from `Extension.install(app)` instead.
