+++ src/extensions/index.ts
@@ -1,10 +1,1 @@
-export type {
-  ApplicationSystemBinding,
-  AssetBinding,
-  AssetHandler,
-  AssetLoadRequest,
-  Extension,
-  ExtensionDisposer,
-  RendererBinding,
-  SerializerBinding,
-} from './Extension';
+export type { AssetBinding, AssetHandler, AssetLoadRequest, Extension, ExtensionDisposer, RendererBinding, SerializerBinding } from './Extension';
