+++ test/core/application-lifecycle.test.ts
@@ -206,7 +206,6 @@ const loadHarness = async (options: LifecycleHarnessOptions = {}): Promise<Lifec
     }),
   }));
   vi.doMock('#extensions/materialize', () => ({
-    materializeApplicationSystems: vi.fn(),
     materializeAssetBindings: options.materializeAssetBindings ?? vi.fn(),
     materializeRendererBindings: options.materializeRendererBindings ?? vi.fn(),
     materializeSerializerBindings: options.materializeSerializerBindings ?? vi.fn(),
