+++ test/core/focus-visibility.test.ts
@@ -105,7 +105,6 @@ const loadHarness = async (): Promise<FocusVisibilityHarness> => {
     }),
   }));
   vi.doMock('#extensions/materialize', () => ({
-    materializeApplicationSystems: vi.fn(),
     materializeAssetBindings: vi.fn(),
     materializeRendererBindings: vi.fn(),
     materializeSerializerBindings: vi.fn(),
