+++ test/rendering/webgl2-backend.test.ts
@@ -142,7 +142,6 @@ describe('Application.setCursor', () => {
       }),
     }));
     vi.doMock('#extensions/materialize', () => ({
-      materializeApplicationSystems: vi.fn(),
       materializeAssetBindings: vi.fn(),
       materializeRendererBindings: vi.fn(),
       materializeSerializerBindings: vi.fn(),
